Start selecting code to hammer at.
From all reports it seems the image integration and labelling takes all
the time that koki spends beating the image. The purpose here is to
clone all the functions that actually do the work, then gently massage
them to be faster. That way, there's always a canonical copy where it's
clear what's happening, and a possibly faster one where it's less clear.

I'm prefixing all the functions I've touched with 'funky'. I've made no
change to the actual bodies of the functions except function call
renames, and adding 'static' to a bunch of functions. Everything except
koki_funky_label_adaptive is static.

Change-Id: I42cc0c76cbec918f838bd050c7d313badd592b98