Only maintain a portion of the integral image.
That is: we only store the last 16 scan lines. It's kept in a single
buffer, and we switch between lines by taking y % 16. That means we're
effectively storing data in a subportion of the integral image and
discarding it after a while: this reduces cache pressure.

I originally had it limited to the window size, however a) there are
some off-by-ones to deal with there, and b) it requires an integer
division / modulus, where using %16 becomes &16, which is faster.

Passes the code recovery tests with exactly the same behaviours.

Change-Id: I1b40b477ba7d2c80457bf281bd78e9a750f824d4