Calculate sampling window on the fly.
That is: repeatedly increment the xpos / ypos and so forth that we're
sampling, and then clamp it to sane boundries. This avoids a function
call, and some nontrivial branching. On x86_64, there's approximately a
7% performance improvement.

My belief is that, on ARM, this'll be translated to conditional
instructions rather than branches, which'll be much more effective.
It'll also reduce the icache pressure (not that there is any).

Change-Id: I6a92e1c080d9dd96bde5b9372a87721fed923a55