Inject some SSE.
This is a test to see what the outcome of applying SSE to libkoki will
be -- i.e. how much speed improvement we get from using SIMD. It'll also
be a useful starting point for using NEON on ARM.

For the moment, this appears to achieve no speed improvement... possibly
because amd64 processors are so massively optimised (they have 4 execute
units by themselves...) and SSE instructions /aren't/ so optimised by
the processor. Or whatever; more workd required.